door_entry_system/docker_container/requirements.txt
RPi.GPIO
time
flask